arch-vega,gpu-compute: Add GPU pseudo instructions
On the CPU side of gem5, magic instructions can be used to implement
instructions that do not exist on real hardware. This is useful for
experimenting adding one-off instructions in a real application, for
example. There is currently no way to do this on the GPU side in gem5,
however.

This patch allows for pseudo instructions by leveraging the unused
S_TRAP instruction. This instruction takes a single 16-bit immediate
value which can be used as the pseudo instruction ID. On real hardware,
only the lower 8 bits of the immediate are used, so we can use the other
bits to mark the immediate as a gem5 pseudo instruction.

The trap is intended to mark the *next* instruction as a pseudo
instruction. This can be used to add a different implementation to an
instruction. For example, we can code a normal multiply instruction into
an application (e.g., V_MUL_F32) and prefix with a trap that indicates
the data type is another type to evaluate support for new data types. By
essentially prefixing instructions, there are many different things that
can be done with any instruction.

The patch does not include example code, but the implementation of the
above example would be implemented in the instruction's execute() method
directly.

Change-Id: I2b2a9072521c9757c48b53c4d57413a7774222b2