gpu-compute,util: GPU Instruction test tracer
This code adds a debug flag used to print *dynamic* traces of
instructions that can be used to verify the implementations of
instructions in gem5. It modifies the GPU dynamic instruction and
register files to output the values of registers before/after
instruction execution, the instruction assembly from gem5, and uses the
instruction flags to determine a test type. A trace is created when the
debug flag is enabled and written to gem5.gitt.gz in m5out by default.
For example:

build/VEGA_X86/gem5.opt --debug-flags=GPUInstTestTrace \
    configs/example/gpufs/vega10_kvm.py <args ...>

This is used in conjunction with a python script that will read the
trace and generate C++ HIP code that runs on a real GPU to compare
hardware output with gem5's output. The python script can generate
multiple tests at once, but is intended to be limited to single
wavefront execution. The python script should therefore specify a single
dynamic wavefront ID followed by a range of instruction sequence numbers
to generate tests for. For example:

./util/gpu_inst_test/multitest.py -i m5out/gem5.gitt.gz -w 2558 -s \
    143400 -e 143600 -o verify.cpp
hipcc verify.cpp -o verify
./verify

Change-Id: I25d74c21cde0297ced0627d4ec8e8b91df8fde2a