arch-vega: WIP: Bytecode-base GPU pseudo instruction
Pseudo instructions are one of the most compelling reasons for
simulation. Currently these are not supported in mainstream gem5. This
commit adds a bytecode-based implementation of GPU pseudo instructions.
This implementation is similar to the CPU pseudo instructions (the non
_addr variant) that emit bytecode directly in the program binary which
can be decoded by the simulators ISA decoder.

An alternate way to implement GPU pseudo instructions is to use an
unused instruction and/or prefix instructions which an unused
instruction such as S_TRAP. The advantages of that approach are that
implementing a new pseudo instruction is as easy as hijacking the
execute method of the instruction being prefixed. The downside is S_TRAP
causes the wavefront to enter a special state that might be performance
limiting compared to bytecode-based pseudo instructions.

On the other hand, bytecode-based instructions are more complicated to
implement in the simulator, as a decoded instruction class and
instruction class must be implemented in the ISA decoder/instruction
files. Once that work is done they are much more streamlined into the
GPU code.

Change-Id: I5f56ce5e08ca5c2ddea40a4ac4ea08817e5233cc