misc: bash scripts to build/clean/run gfx902 SE mode
This tests most of the weekly/nightly SE mode deprecated tests on the
gfx902 target. It is not intended to ever be pushed to mainline gem5. It
is intended as a reference point to folk would like to reproduce the
work.

The most notable missing tests are the most simple ones: square and
hip-samples. If these tests work those probably do too, but I will add
those at some point as well.